% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_xlsx_file_and_sheet}
\alias{.check_xlsx_file_and_sheet}
\title{Sanitize input xlsx file}
\usage{
.check_xlsx_file_and_sheet(file, required_sheets)
}
\arguments{
\item{file}{A character string with a path pointing to an ARM xlsx export file}

\item{required_sheets}{A vector of character strings that must be available in the xlsx export file}
}
\description{
Ensure that file format and sheets match expectations
}
