% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_description.R
\name{parse_description}
\alias{parse_description}
\title{Extract and organize trial metadata}
\usage{
parse_description(descr_sheet)
}
\arguments{
\item{descr_df}{Tibble as output by \code{read_descr_sheet}; must contain all standard ARM description fields.}
}
\value{
Named list with elements for sowing, site, plots, etc.
}
\description{
Parses high-level study and management information from a raw description tibble, returning a structured list for later formatting.
}
