% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_description.R
\name{parse_maintenance}
\alias{parse_maintenance}
\title{Parse trial maintenance activities}
\usage{
parse_maintenance(descr_sheet)
}
\arguments{
\item{descr_df}{Tibble with trial metadata, as from \code{read_descr_sheet}.}
}
\value{
Data frame or list of parsed maintenance events, with dates and categories.
}
\description{
Extracts information on maintenance events and management practices from the description data. Intended for generating treatment/operation tables.
}
