% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_opviarm_app.R
\name{run_opviarm_app}
\alias{run_opviarm_app}
\title{Launch the OpviARM Shiny App}
\usage{
run_opviarm_app()
}
\value{
No return value. Called for side effects.
}
\description{
Starts the web interface for uploading and parsing ARM export files and sending data to Opvia.
The app allows users to upload assessment, description, treatment, and field map files, preview their contents, and (optionally) submit the formatted data to Opvia.
}
