% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_to_opvia.R
\name{update_table}
\alias{update_table}
\title{Add or update records in an Opvia table}
\usage{
update_table(formatted_df, tableID)
}
\arguments{
\item{tableID}{UUID string for the target Opvia table.}

\item{data}{Data frame with rows and columns matching the Opvia table schema.}
}
\value{
Logical value indicating whether the upload/update succeeded.
}
\description{
Pushes new records or updates existing entries in an Opvia table using data from an R data frame.
}
